Update GUI.java 1.7
Updated the GUI to give a message to the user when an order cannot be completed, with the food shortages that caused the order to not be completed. As well, 
make sure that if a order fails halfway through being completed. (like if hamper 1 goes through, but fails on hamper 2) that the AvailableFood object is reset, because before this would result in the food from the first hamper being taken, even if the overall order failed.